move run to docker_run and run_manager to run
